Feat: Navigate to Home after successful authentication
This commit updates the `NavGraph.kt` to navigate from `AuthScreen` to `HomeGraph` upon successful authentication.

The `AuthScreen` composable now accepts a lambda function that is invoked when authentication is successful. This lambda navigates to `Screen.HomeGraph` and clears the back stack up to `Screen.Auth` to prevent users from navigating back to the authentication screen.